+++ R/pointdensity.R
@@ -1,0 +1,45 @@
+#######################################################################
+# dbscan - Density Based Clustering of Applications with Noise
+#          and Related Algorithms
+# Copyright (C) 2017 Michael Hahsler
+
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with this program; if not, write to the Free Software Foundation, Inc.,
+# 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+
+pointdensity <- function(x, eps, type = "frequency",
+  search = "kdtree", bucketSize = 10,
+  splitRule = "suggest", approx = 0){
+
+  ### ANN uses squared distance
+  eps2 <- eps*eps
+
+  type <- match.arg(type, choices = c("frequency", "density"))
+
+  search <- pmatch(toupper(search), c("KDTREE", "LINEAR", "DIST"))
+  if(is.na(search)) stop("Unknown NN search type!")
+
+  splitRule <- pmatch(toupper(splitRule), .ANNsplitRule)-1L
+  if(is.na(splitRule)) stop("Unknown splitRule!")
+
+  d <- dbscan_density_int(as.matrix(x), as.double(eps2),
+  as.integer(search), as.integer(bucketSize),
+  as.integer(splitRule), as.double(approx))
+
+  if(type == "density") d <- d / (2*eps*nrow(x))
+
+  d
+}
+
+